+++ atomate/vasp/analysis/linear_response.py
@@ -253,29 +253,32 @@ def inverse_matrix_uncertainty(matrix, matrix_covar):
     jacobians = [[] for i in range(m)]
 
     det = np.linalg.det(matrix)
-    for i in range(m):
-        for j in range(n):
-            mji = np.delete(np.delete(matrix, j, 0), i, 1)
-            minor = (-1) ** (i + j) * np.linalg.det(mji)
+    for idx in range(m):
+        for jdx in range(n):
+            mji = np.delete(np.delete(matrix, jdx, 0), idx, 1)
+            minor = (-1) ** (idx + jdx) * np.linalg.det(mji)
 
             j_matrix = np.zeros([m, n])
-            for k in range(m):
-                for l in range(n):
-                    det_p = det_deriv(matrix, k, l)
+            for kdx in range(m):
+                for ldx in range(n):
+                    det_p = det_deriv(matrix, kdx, ldx)
 
-                    if k == j or l == i:
+                    if kdx == jdx or ldx == idx:
                         minor_p = 0.0
                     else:
-                        kk, ll = k - 1 if k > j else k, l - 1 if l > i else l
-                        minor_p = (-1) ** (i + j) * det_deriv(mji, kk, ll)
+                        kk, ll = (
+                            kdx - 1 if kdx > jdx else kdx,
+                            ldx - 1 if ldx > idx else ldx,
+                        )
+                        minor_p = (-1) ** (idx + jdx) * det_deriv(mji, kk, ll)
 
-                    j_matrix[k, l] = (minor_p * det - minor * det_p) / det ** 2
+                    j_matrix[kdx, ldx] = (minor_p * det - minor * det_p) / det**2
 
-            jacobians[i].append(j_matrix)
+            jacobians[idx].append(j_matrix)
 
             j_vec = np.reshape(j_matrix, [m * n, 1])
             sigma_f = np.sum(np.dot(np.transpose(j_vec), np.dot(matrix_covar, j_vec)))
-            matrixinv_var[i, j] = sigma_f
+            matrixinv_var[idx, jdx] = sigma_f
 
     return matrixinv, matrixinv_var, jacobians
 
@@ -312,7 +315,7 @@ def chi_inverse(chi, chi_err, method="full"):
         )
 
     # Assume cross-covariances are zero
-    chi_covar = np.diag(np.reshape(chi_err_block ** 2, [n_response * n_response]))
+    chi_covar = np.diag(np.reshape(chi_err_block**2, [n_response * n_response]))
 
     (chi_inv, chi_inv_var, chi_inv_jacobs) = inverse_matrix_uncertainty(
         chi_block, chi_covar
@@ -359,12 +362,12 @@ def compute_uj_simple_two_by_two(
     umat_err = f_matrix_err[2 * i : 2 * (i + 1), 2 * i : 2 * (i + 1)]
 
     uval = 0.25 * np.sum(umat)
-    uval_err = 0.25 * np.sqrt(np.sum(umat_err ** 2))
+    uval_err = 0.25 * np.sqrt(np.sum(umat_err**2))
 
     jmat = np.array([[-1, 1], [1, -1]]) * umat.copy()
     jmat_err = umat_err.copy()
     jval = 0.25 * np.sum(jmat)
-    jval_err = 0.25 * np.sqrt(np.sum(jmat_err ** 2))
+    jval_err = 0.25 * np.sqrt(np.sum(jmat_err**2))
 
     return uval, uval_err, jval, jval_err
 
@@ -439,7 +442,7 @@ def compute_uj_scaled_two_by_two(
     uval_err = uval_err + np.sum(
         np.dot(
             np.transpose(jacob_vec),
-            np.dot(np.diag(np.reshape(chi_sub_scf_err ** 2, [4])), jacob_vec),
+            np.dot(np.diag(np.reshape(chi_sub_scf_err**2, [4])), jacob_vec),
         )
     )
     # nscf component
@@ -460,7 +463,7 @@ def compute_uj_scaled_two_by_two(
     uval_err = uval_err + np.sum(
         np.dot(
             np.transpose(jacob_vec),
-            np.dot(np.diag(np.reshape(chi_sub_nscf_err ** 2, [4])), jacob_vec),
+            np.dot(np.diag(np.reshape(chi_sub_nscf_err**2, [4])), jacob_vec),
         )
     )
     # compute std
@@ -501,7 +504,7 @@ def compute_uj_scaled_two_by_two(
     jval_err = jval_err + np.sum(
         np.dot(
             np.transpose(jacob_vec),
-            np.dot(np.diag(np.reshape(chi_sub_scf_err ** 2, [4])), jacob_vec),
+            np.dot(np.diag(np.reshape(chi_sub_scf_err**2, [4])), jacob_vec),
         )
     )
     # nscf component
@@ -522,7 +525,7 @@ def compute_uj_scaled_two_by_two(
     jval_err = jval_err + np.sum(
         np.dot(
             np.transpose(jacob_vec),
-            np.dot(np.diag(np.reshape(chi_sub_nscf_err ** 2, [4])), jacob_vec),
+            np.dot(np.diag(np.reshape(chi_sub_nscf_err**2, [4])), jacob_vec),
         )
     )
     # compute std