+++ atomate/vasp/drones.py
@@ -11,13 +11,12 @@ import os
 import re
 import traceback
 import warnings
-from collections import OrderedDict
 from fnmatch import fnmatch
+from shutil import which
 
 import numpy as np
 from monty.io import zopen
 from monty.json import jsanitize
-from monty.os.path import which
 from pymatgen.apps.borg.hive import AbstractDrone
 from pymatgen.command_line.bader_caller import bader_analysis_from_path
 from pymatgen.core.composition import Composition
@@ -235,10 +234,10 @@ class VaspDrone(AbstractDrone):
             file_pattern (string): files to be searched for
 
         Returns:
-            OrderedDict of the names of the files to be processed further.
-            The key is set from list of run types: self.runs
+            dict: names of the files to be processed further. The key is set from list
+                of run types: self.runs
         """
-        processed_files = OrderedDict()
+        processed_files = {}
         files = os.listdir(path)
         for r in self.runs:
             # try subfolder schema
@@ -420,6 +419,11 @@ class VaspDrone(AbstractDrone):
                 for k in ["optical_absorption_coeff", "dielectric"]:
                     d["output"][k] = d_calc_final["output"][k]
 
+            # store optical data, overwrites the LOPTICS data
+            if d["input"]["incar"].get("ALGO") == "CHI":
+                for k in ["optical_absorption_coeff", "dielectric"]:
+                    d["output"][k] = d_calc_final["output"][k]
+
             d["state"] = (
                 "successful" if d_calc["has_vasp_completed"] else "unsuccessful"
             )
@@ -577,6 +581,14 @@ class VaspDrone(AbstractDrone):
 
         # parse output from loptics
         if vrun.incar.get("LOPTICS", False):
+            dielectric = vrun.dielectric
+            d["output"]["dielectric"] = dict(
+                energy=dielectric[0], real=dielectric[1], imag=dielectric[2]
+            )
+            d["output"]["optical_absorption_coeff"] = vrun.optical_absorption_coeff
+
+        # parse output from response function
+        if vrun.incar.get("ALGO") == "CHI":
             dielectric = vrun.dielectric
             d["output"]["dielectric"] = dict(
                 energy=dielectric[0], real=dielectric[1], imag=dielectric[2]
@@ -816,9 +828,9 @@ class VaspDrone(AbstractDrone):
         # notice if the validation fails? -computron
         for k, v in self.schema.items():
             if k == "calcs_reversed":
-                diff = v.difference(set(d.get(k, d)[0].keys()))
+                diff = v - set(d.get(k, d)[0].keys())
             else:
-                diff = v.difference(set(d.get(k, d).keys()))
+                diff = v - set(d.get(k, d).keys())
             if diff:
                 logger.warning(f"The keys {diff} in {k} not set")