+++ atomate/vasp/firetasks/glue_tasks.py
@@ -290,15 +290,17 @@ class GetInterpolatedPOSCAR(FiretaskBase):
         # use CopyFilesFromCalcLoc to get files from previous locations.
         CopyFilesFromCalcLoc(
             calc_loc=self["start"],
-            filenames=["CONTCAR"],
+            filenames=["CONTCAR", "CONTCAR.gz"],
             name_prepend=interpolate_folder + os.sep,
             name_append="_0",
+            decompress=True,
         ).run_task(fw_spec=fw_spec)
         CopyFilesFromCalcLoc(
             calc_loc=self["end"],
-            filenames=["CONTCAR"],
+            filenames=["CONTCAR", "CONTCAR.gz"],
             name_prepend=interpolate_folder + os.sep,
             name_append="_1",
+            decompress=True,
         ).run_task(fw_spec=fw_spec)
 
         # assuming first calc_dir is polar structure for ferroelectric search