+++ atomate/vasp/firetasks/write_inputs.py
@@ -630,7 +630,6 @@ class WriteTransmutedStructureIOSet(FiretaskBase):
             t_cls = None
             for m in [
                 "advanced_transformations",
-                "defect_transformations",
                 "site_transformations",
                 "standard_transformations",
             ]: