+++ atomate/vasp/fireworks/core.py
@@ -2,7 +2,10 @@
 Defines standardized Fireworks that can be chained easily to perform various
 sequences of VASP calculations.
 """
+from __future__ import annotations
+
 import warnings
+from typing import Any
 
 from fireworks import Firework
 from pymatgen.core import Structure
@@ -583,26 +586,26 @@ class NonSCFFW(Firework):
 class DFPTFW(Firework):
     def __init__(
         self,
-        structure=None,
-        prev_calc_dir=None,
-        name="static dielectric",
-        vasp_cmd=VASP_CMD,
-        copy_vasp_outputs=True,
-        lepsilon=True,
-        db_file=DB_FILE,
-        parents=None,
-        user_incar_settings=None,
-        pass_nm_results=False,
+        structure: Structure = None,
+        prev_calc_dir: str = None,
+        name: str = "static dielectric",
+        vasp_cmd: str = VASP_CMD,
+        copy_vasp_outputs: bool = True,
+        lepsilon: bool = True,
+        db_file: str = DB_FILE,
+        parents: Firework | list[Firework] = None,
+        user_incar_settings: dict[str, Any] = None,
+        pass_nm_results: bool = False,
         **kwargs,
-    ):
+    ) -> None:
         """
-         Static DFPT calculation Firework
+        Static DFPT calculation Firework
 
         Args:
             structure (Structure): Input structure. If copy_vasp_outputs, used only to set the
                 name of the FW.
             name (str): Name for the Firework.
-            lepsilon (bool): Turn on LEPSILON to calculate polar properties
+            lepsilon (bool): Turn on LEPSILON to calculate polar properties. Defaults to True.
             vasp_cmd (str): Command to run vasp.
             copy_vasp_outputs (str or bool): Whether to copy outputs from previous
                 run. Defaults to True.