+++ atomate/vasp/workflows/tests/test_exchange_workflow.py
@@ -1,10 +1,10 @@
 import os
 import unittest
+from shutil import which
 
 import pandas as pd
 from fireworks import FWorker
 from fireworks.core.rocket_launcher import rapidfire
-from monty.os.path import which
 from pymatgen.core import Structure
 
 from atomate.utils.testing import AtomateTest